refactor(domain): turn script into an immutable value type
Merge script_basis into script as a single simple value type. Drop the
mutable cache of parsed operations, the upgrade mutex, and the lazy
re-parse logic. operations() now returns operation::list by value, so
callers in hot paths must hoist the call outside the loop; caching is
the caller's responsibility.

Net diff: -1024 LOC after removing script_basis.hpp/.cpp and the
duplicated static helpers between the two classes.

Iteration helpers (begin/end/size/front/back/empty/clear/operator[])
were removed because they depended on the internal cache. Callers were
updated to either call operations() once and iterate the returned
list, or to operate on bytes() directly for "is the script empty"
checks. The C-API was regenerated to drop the corresponding free
functions; tests were updated accordingly.

Follows the same philosophy as #364 (header): domain classes stay
simple — no internal caches, mutable state, or optimisations.
