+++ src/domain/src/machine/program.cpp
@@ -43,26 +43,29 @@ void program::reserve_stacks() {
 
 program::program()
     : script_(&default_script_),
+      script_ops_(chain::operations(default_script_)),
       transaction_(&default_tx_),
-      jump_(default_script_.begin())
+      jump_(script_ops_.begin())
 {
     reserve_stacks();
 }
 
 program::program(script const& script)
     : script_(&script),
+      script_ops_(chain::operations(script)),
       transaction_(&default_tx_),
-      jump_(script.begin()) {
+      jump_(script_ops_.begin()) {
     reserve_stacks();
 }
 
 program::program(script const& script, chain::transaction const& transaction, uint32_t input_index, script_flags_t flags, uint64_t value)
     : script_(&script),
+      script_ops_(chain::operations(script)),
       transaction_(&transaction),
       input_index_(input_index),
       flags_(flags),
       value_(value),
-      jump_(script.begin()),
+      jump_(script_ops_.begin()),
       context_(script_execution_context(input_index, transaction)) {
     reserve_stacks();
 }
@@ -80,6 +83,7 @@ program::program(
 #endif // ! KTH_CURRENCY_BCH
 )
     : script_(&script),
+      script_ops_(chain::operations(script)),
       transaction_(&transaction),
       input_index_(input_index),
       flags_(flags),
@@ -87,7 +91,7 @@ program::program(
 #if ! defined(KTH_CURRENCY_BCH)
       version_(version),
 #endif // ! KTH_CURRENCY_BCH
-      jump_(script.begin()),
+      jump_(script_ops_.begin()),
       primary_(std::move(stack)),
       context_(script_execution_context(input_index, transaction)) {
     reserve_stacks();
@@ -97,6 +101,7 @@ program::program(
 // Metrics are propagated so sig_checks accumulate across evaluations (matching BCHN).
 program::program(script const& script, program const& x)
     : script_(&script),
+      script_ops_(chain::operations(script)),
       transaction_(x.transaction_),
       input_index_(x.input_index_),
       flags_(x.flags_),
@@ -104,7 +109,7 @@ program::program(script const& script, program const& x)
 #if ! defined(KTH_CURRENCY_BCH)
       version_(x.version_),
 #endif // ! KTH_CURRENCY_BCH
-      jump_(script.begin()),
+      jump_(script_ops_.begin()),
       primary_(x.primary_),
       metrics_(x.metrics_),
       context_(x.context_) {
@@ -115,6 +120,7 @@ program::program(script const& script, program const& x)
 // Metrics are propagated so sig_checks accumulate across evaluations (matching BCHN).
 program::program(script const& script, program&& x, bool /*unused*/)
     : script_(&script),
+      script_ops_(chain::operations(script)),
       transaction_(x.transaction_),
       input_index_(x.input_index_),
       flags_(x.flags_),
@@ -122,7 +128,7 @@ program::program(script const& script, program&& x, bool /*unused*/)
 #if ! defined(KTH_CURRENCY_BCH)
       version_(x.version_),
 #endif // ! KTH_CURRENCY_BCH
-      jump_(script.begin()),
+      jump_(script_ops_.begin()),
       primary_(std::move(x.primary_)),
       metrics_(x.metrics_),
       context_(std::move(x.context_)) {
